fix : List - control bar layout
