+++ src/views/List/StockList.tsx
@@ -1,15 +1,12 @@
-import { Suspense, lazy, useEffect, useState } from 'react';
+import { Suspense, useEffect, useState } from 'react';
 import styled from 'styled-components';
 
 import { BaseInput } from '../../components/Input/BaseInput';
-import GroupSummary from './GroupSummary/GroupSummary';
 import useSaveChangedStocksData from './hooks/useSaveChangedStocksData';
 import useSaveChangedGroupsData from './hooks/useSaveChangedGroupedData';
 import StockListSkeleton from './StockListSkeleton';
 import useSaveChangedSoldsData from './hooks/useSaveChangedSoldData';
 import StockListContent from './StockListContent';
-
-const GroupButtons = lazy(() => import('./GroupButtons/GroupButtons'));
 
 const StockList = () => {
   const [firstLoad, setFirstLoad] = useState(true);
@@ -25,10 +22,6 @@ const StockList = () => {
   return (
     <Suspense fallback={<StockListSkeleton />}>
       <StyledStockList>
-        <StyledControlBar>
-          <GroupSummary />
-          <GroupButtons />
-        </StyledControlBar>
         <StockListContent />
         <div className='container'></div>
       </StyledStockList>
@@ -47,16 +40,5 @@ const StyledStockList = styled('div')`
       -webkit-text-fill-color: ${({ theme }) => theme.colors.grey900};
       opacity: 1;
     }
-  }
-`;
-
-const StyledControlBar = styled('div')`
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap-reverse;
-  row-gap: 10px;
-  padding: 20px 0 10px 0;
-
-  @media ${({ theme }) => theme.devices.laptop} {
   }
 `;
