+++ src/views/List/StockListContent/StockTableMenu/EditCurrentPrice/index.tsx
@@ -33,7 +33,7 @@ const StyledText = styled.span`
   margin-left: 5px;
   color: inherit;
 
-  @media ${({ theme }) => theme.devices.mobile} {
+  @media ${({ theme }) => theme.devices.laptop} {
     display: none;
   }
 `;
