+++ src/views/List/StockListSkeleton.tsx
@@ -8,9 +8,7 @@ const StockListSkeleton = () => {
   return (
     <div>
       <StyledControlBar>
-        <div>
-          <GroupButtonSkeleton />
-        </div>
+        <GroupButtonSkeleton />
       </StyledControlBar>
       <GroupSummarySkeleton />
       <ListTableSkeleton />
@@ -24,6 +22,7 @@ export default StockListSkeleton;
 const StyledControlBar = styled('div')`
   display: flex;
   justify-content: space-between;
+  padding: 20px 0 20px 0;
 
   @media ${({ theme }) => theme.devices.tablet} {
     flex-direction: column;
